PulseOxicometer MAX30100 code.
It needs to be a little modified before using it with arduino.
Find more at the link inside the code.(wrong SDC,SCL voltage levels, and some resistors need to be taken out and put anohers, or if you have the green module, you can cut a trace and put a wire for it to work).